Use fixed-point math to scale mouse movement.
Improves virtual mouse performance from 'awful' to 'ok,' still needs work.

+++ firmware/drivers/virtual.h
@@ -12,7 +12,7 @@
 #include <stdbool.h>
 
 typedef struct {
-	int32_t x, y;
+	int16_t x, y;
 	uint8_t buttons;
 } virtual_mouse_data;
